Rework observer to log before and after epochs
Why:

The current way logging was done in classification task was confusing
and lead to bugs. The accuracy reported at end of epoch was actually
the one before the epoch, and accuracy of last epoch was not reported at
all. The order of metrics was also a cause for this issue, because the
Accuracy metric was called after ProgressView, causing a shift between
actual epoch metrics and what was printed.