Seed options for glue (#107)
* added option to sed seed

* added option to redirect log

* Rework observer to log before and after epochs

Why:

The current way logging was done in classification task was confusing
and lead to bugs. The accuracy reported at end of epoch was actually
the one before the epoch, and accuracy of last epoch was not reported at
all. The order of metrics was also a cause for this issue, because the
Accuracy metric was called after ProgressView, causing a shift between
actual epoch metrics and what was printed.

* Add warmup schedule for BERT-GLUE

* switched to max length 512

* now can set model cache folder

* added option to specify the tokenizer cache

* back to 128 tokens

* better log

* avoiding resetting LR decay at every epoch

* added option for weight decay

* passing model seed

* correctly passing the cache

* added model seed

* fixed typo

* more typos..

* model_seed=>global_seed / fixed duplicate line

* renamed env variables

* fixed bug in classification

* Remove model_seed from bert (now using global)

* Use option() to get cache dirs

Why:

The standard way to get such options is with option(), which will
fetch anyway env vars values for OLYMPUS_{THE_KEY}. Using option()
also gives the possibility to update the value from within python
without touching env vars.

* fixed return

* priting seeds

* moved print intside main function

* removed old script

Co-authored-by: Xavier Bouthillier <xavier.bouthillier@umontreal.ca>