Fix unsynced hpo duplicates
Why:

When there is accidentally duplicate HPOs, some of them may have
suggested more trials yet. When fixing the duplicates, we simply retain
the last one and stop the others (looking for the most advanced one
would be more complex and expansive operations on the DB...). So the
HPO with less duplicates will raise an error when observing trials it
did not sample yet.

How:

Just ignore the unknown trials until they are suggested. This will lead
to to duplicate trials but it's fine as long as the checkpointing is
used, so the duplicated one will just reload the checkpoint and return
final results.